Revert "Merge pull request #6 from jvanmelckebeke/master"
This reverts commit 9e8c863603d64b0fbbaab573532c4650fc1d579e, reversing
changes made to c6708a12d1987e23cac8bc6ae58e6d9000a34f3b.

+++ src/be/ac/kuleuven/cs/drama/vertalerpack/vertaler/Vertaler2.java
@@ -74,7 +74,7 @@ public class Vertaler2 {
 
    public int translateLine(int line) {
       try {
-         return (Integer) _linemap.get(line - 1);
+         return ((Integer) _linemap.get(line - 1)).intValue();
       } catch (Exception e) {
          System.out.println("*** " + e);
          return -1;
